+++ src/_toPath.ts
@@ -4,16 +4,17 @@ function toPath(path: Array<string | number> | string | number): string[] {
   let chain = path
   if (isArray(chain)) {
     chain = chain.join('.')
-  } else {
-    chain += ''
   }
 
-  const rs = (chain + '')
-    .replace(/\[([^\]]+)\]/gm, '.$1')
-    .replace(/^\./g, '')
-    .split('.')
+  let rs = chain + ''
+  if (rs.includes('[')) {
+    rs = rs.replace(/\[(['"])?([^\]'"]+)\1?\]/gm, '.$2')
+  }
+  if (rs[0] === '.') {
+    rs = rs.substring(1)
+  }
 
-  return rs
+  return rs.split('.')
 }
 
 export default toPath