+++ src/math/min.ts
@@ -1,5 +1,7 @@
-import flatMap from '../collection/flatMap'
-import isNil from '../is/isNil'
+import isArray from "../is/isArray";
+import isNumeric from "../is/isNumeric";
+import isSet from "../is/isSet";
+
 /**
  * 返回给定数字序列中最小的一个。忽略NaN，null，undefined
  * @example
@@ -10,16 +12,20 @@ import isNil from '../is/isNil'
  * //-Infinity
  * console.log(_.min([-Infinity,-9999,0,null]))
  * @param values 数字/字符数组/Set
- * @returns
+ * @returns 如果参数不是数组/Set，返回NaN
  * @since 1.0.0
  */
 function min(
   values: Set<string | number> | Array<string | number>
 ): number {
-  const vals = flatMap<any>(values, v => isNil(v) || isNaN(v) ? [] : v)
-  let f64a = new Float64Array(vals)
-  f64a.sort()
-  return f64a[0]
+  if (!isArray(values) && !isSet(values)) return NaN
+  let rs: number | string = isArray(values) ? values[0] : values.values().next().value!
+  values.forEach(v => {
+    if (isNumeric(v) && v < rs) {
+      rs = v
+    }
+  });
+  return Number(rs)
 }
 
 export default min