+++ src/string/escapeRegExp.ts
@@ -1,20 +1,5 @@
-import toString from "./toString"
-const REG_EXP_KEYWORDS: string[] = [
-  '\\',
-  '$',
-  '(',
-  ')',
-  '*',
-  '+',
-  '.',
-  '[',
-  ']',
-  '?',
-  '^',
-  '{',
-  '}',
-  '|',
-]
+import toString from "./toString";
+
 /**
  * 转义正则字符串中的特殊字符，包括 '\', '$', '(', ')', '*', '+', '.', '[', ']', '?', '^', '\{', '\}', '|'
  *
@@ -27,9 +12,24 @@ const REG_EXP_KEYWORDS: string[] = [
  * @since 1.0.0
  */
 function escapeRegExp(str: any): string {
-  return toString(str)
-    .split('')
-    .reduce((a, b) => a + (REG_EXP_KEYWORDS.includes(b) ? '\\' + b : b), '')
+  let rs = "";
+  str = toString(str)
+  for (let i = 0; i < str.length; i++) {
+    let s = str[i];
+    const code = s.charCodeAt(0);
+    if (
+      code === 36 ||
+      code === 46 ||
+      code === 63 ||
+      (code >= 40 && code <= 43) ||
+      (code >= 91 && code <= 94) ||
+      (code >= 123 && code <= 125)
+    ) {
+      s = "\\" + s;
+    }
+    rs += s;
+  }
+  return rs;
 }
 
 export default escapeRegExp